requirements.txt
geopandas==1.0.1
setuptools==75.6.0
pandas==2.2.3
numpy==1.26.4
SQLAlchemy==2.0.36
psycopg2_binary==2.9.10
sqlparse==0.5.3
scipy==1.13.1
pandapower==2.14.11
openpyxl==3.1.5
xlrd==2.0.1
python-dotenv==1.0.1
matplotlib==3.10.0
plotly==5.24.1
seaborn==0.13.2
contextily==1.6.2
igraph==0.11.8
sphinx-autoapi==3.4.0
sphinx-js==4.0.0
sphinx-rtd-theme==3.0.2